Implement native roll_bits and roll_bytes

+++ bhv/np.py
@@ -287,12 +287,13 @@ class NumPyPacked64BHV(AbstractBHV):
 
     def roll_word_bits(self, n: int) -> 'NumPyPacked64BHV':
         assert abs(n) < 64, "only supports 64 rolls"
+        # https://github.com/numba/numba/issues/6381
         if n == 0:
             return NumPyPacked64BHV(self.data)
         elif n > 0:
-            return NumPyPacked64BHV(np.bitwise_or(np.right_shift(self.data, n), np.left_shift(self.data, (64 - n))))
+            return NumPyPacked64BHV(np.bitwise_or(np.right_shift(self.data, np.uint64(n)), np.left_shift(self.data, np.uint64(64 - n))))
         else:
-            return NumPyPacked64BHV(np.bitwise_or(np.left_shift(self.data, n), np.right_shift(self.data, (64 - n))))
+            return NumPyPacked64BHV(np.bitwise_or(np.left_shift(self.data, np.uint64(n)), np.right_shift(self.data, np.uint64(64 - n))))
 
     # roll_words and roll_word_bits could be combined for more options allowing positive and negative combinations
     # ((1 2 3 4) (a b c d) (α β γ δ))
